Update livekit client-sdk-js
We use a git repo directly as a dependency here. In our package.json we
list the repo, but in the yarn.lock there's a specific commit hash that
we install/use.

In order to create this update you can run the following in this repo:
$ yarn add 'git+https://github.com/Invisv-Privacy/client-sdk-js.git'

Then ensure that the change sets the yarn.lock commit hash to the
desired one.